Req (req v0.1.0-dev)

A work-in-progress HTTP client.

Features

Extensibility via request, response, and error steps

Automatic body decompression (via decompress/2 step)

Automatic body decoding (via decode/2 step)

Retries on errors (via retry/2 step)

High-level API

Req.get!("https://api.github.com/repos/elixir-lang/elixir").body["description"]
#=> "Elixir is a dynamic, functional language designed for building scalable and maintainable applications"

Low-level API

Under the hood, Req works by passing a request through a series of steps.

The request struct, %Req.Request{}, initially contains data like HTTP method and request headers. You can also add request, response, and error steps to it.

Request steps are used to refine the data that will be sent to the server.

After making the actual HTTP request, we'll either get a HTTP response or an error. The request, along with the response or error, will go through response or error steps, respectively.

Nothing is actually executed until we run the pipeline with Req.run/1.

Example:

Req.build(:get, "https://api.github.com/repos/elixir-lang/elixir")
|> Req.add_request_steps([
  &Req.default_headers/1
])
|> Req.add_response_steps([
  &Req.decode/2
])
|> Req.run()
#=> {:ok, %{body: %{"description" => "Elixir is a dynamic," <> ...}, ...}, ...}

We can also build more complex flows like returning a response from a request step or an error from a response step. We will explore those next.

Request steps

A request step is a function that accepts a request and returns one of the following:

A request

A {request, response_or_error} tuple. In that case no further request steps are executed and the return value goes through response or error steps

Examples:

def default_headers(request) do
  update_in(request.headers, &[{"user-agent", "req/0.1.0-dev"} | &1])
end

def read_from_cache(request) do
  case ResponseCache.fetch(request) do
    {:ok, response} -> {request, response}
    :error -> request
  end
end

Response and error steps

A response step is a function that accepts a request and a response and returns one of the following:

A {request, response} tuple

A {request, exception} tuple. In that case, no further response steps are executed but the exception goes through error steps

Similarly, an error step is a function that accepts a request and an exception and returns one of the following:

A {request, exception} tuple

A {request, response} tuple. In that case, no further error steps are executed but the response goes through response steps

Examples:

def decode(request, response) do
  case List.keyfind(response.headers, "content-type", 0) do
    {_, "application/json" <> _} ->
      {request, update_in(response.body, &Jason.decode!/1)}

    _ ->
      {request, response}
  end
end

def log_error(request, exception) do
  Logger.error(["#{request.method} #{request.url}: ", Exception.message(exception)])
  {request, exception}
end

Halting

Any step can call Req.Request.halt/1 to halt the pipeline. This will prevent any further steps from being invoked.

Examples:

def circuit_breaker(request) do
  if CircuitBreaker.open?() do
    {Req.Request.halt(request), RuntimeError.exception("circuit breaker is open")}
  else
    request
  end
end

High-level API

get!(url, opts \\ [])

Makes a GET request.

request(method, url, opts \\ [])

Makes an HTTP request.

Low-level API

add_default_steps(request)

Adds steps that should be reasonable defaults for most users.

request:

default_headers/1

response:

decompress/2

decode/2

add_error_steps(request, steps)

Adds error steps.

add_request_steps(request, steps)

Adds request steps.

add_response_steps(request, steps)

Adds response steps.

build(method, url, opts \\ [])

Builds a request pipeline.

run(request)

Runs a request pipeline.

Returns {:ok, response} or {:error, exception}.

run!(request)

Runs a request pipeline and returns a response or raises an error.

See run/1.

Request steps

default_headers(request)

Adds common request headers.

Currently the following headers are added:

"user-agent" - "req/0.1.0-dev"

Response steps

decode(request, response)

Decodes the response body based on the content-type header.

decompress(request, response)

Decompresses the response body based on the content-encoding header.

Error steps

retry(request, response)

Retries a request in face of errors.

This function can be used as either or both response and error step. It retries a request that resulted in:

a response with status 5xx

an exception