Req.Conn (req v0.6.0-dev)

The Req connection struct.

Types

t()
(since 0.6.0)

@type t() :: %Req.Conn{
  adapter: :finch | :mint | :httpc | :plug | module(),
  assigns: map(),
  halted: boolean(),
  host: String.t() | nil,
  method: String.t(),
  port: non_neg_integer() | nil,
  private: map(),
  query_string: String.t(),
  req_body: iodata() | nil,
  req_headers: Req.Conn.Headers.t(),
  request_path: String.t(),
  resp_body: term(),
  resp_headers: Req.Conn.Headers.t(),
  scheme: :http | :https | nil,
  status: non_neg_integer() | nil
}

Headers

delete_req_header(conn, name)
(since 0.6.0)

@spec delete_req_header(t(), String.t() | atom()) :: t()

Deletes all values of the request header name.

delete_resp_header(conn, name)
(since 0.6.0)

@spec delete_resp_header(t(), String.t() | atom()) :: t()

Deletes all values of the response header name.

drop_req_headers(conn, names)
(since 0.6.0)

@spec drop_req_headers(t(), [String.t() | atom()]) :: t()

Drops all values of the listed request headers.

drop_resp_headers(conn, names)
(since 0.6.0)

@spec drop_resp_headers(t(), [String.t() | atom()]) :: t()

Drops all values of the listed response headers.

put_new_req_header(conn, name, value)
(since 0.6.0)

@spec put_new_req_header(t(), String.t() | atom(), String.t()) :: t()

Adds a request header only if name is not already set.

put_new_resp_header(conn, name, value)
(since 0.6.0)

@spec put_new_resp_header(t(), String.t() | atom(), String.t()) :: t()

Adds a response header only if name is not already set.

put_req_header(conn, name, value)
(since 0.6.0)

@spec put_req_header(t(), String.t() | atom(), String.t()) :: t()

Adds a request header. Multiple values for the same name are kept.

put_resp_header(conn, name, value)
(since 0.6.0)

@spec put_resp_header(t(), String.t() | atom(), String.t()) :: t()

Adds a response header. Multiple values for the same name are kept.

replace_req_header(conn, name, value)
(since 0.6.0)

@spec replace_req_header(t(), String.t() | atom(), String.t()) :: t()

Replaces all values of the request header name with value.

replace_resp_header(conn, name, value)
(since 0.6.0)

@spec replace_resp_header(t(), String.t() | atom(), String.t()) :: t()

Replaces all values of the response header name with value.

req_header_value(conn, name)
(since 0.6.0)

@spec req_header_value(t(), String.t() | atom()) :: String.t() | nil

Returns the first value of the request header name, or nil if absent.

resp_header_value(conn, name)
(since 0.6.0)

@spec resp_header_value(t(), String.t() | atom()) :: String.t() | nil

Returns the first value of the response header name, or nil if absent.

update_req_header(conn, name, default, fun)
(since 0.6.0)

@spec update_req_header(
  t(),
  String.t() | atom(),
  String.t(),
  (String.t() -> String.t())
) :: t()

Updates the existing value of the request header name with fun, inserting default when not present.

update_resp_header(conn, name, default, fun)
(since 0.6.0)

@spec update_resp_header(
  t(),
  String.t() | atom(),
  String.t(),
  (String.t() -> String.t())
) :: t()

Updates the existing value of the response header name with fun, inserting default when not present.

Assigns

assign(conn, assigns)
(since 0.6.0)

@spec assign(t(), assigns :: map() | keyword()) :: t()

Assigns multiple values to the conn.

Examples

iex> conn = Req.conn("http://example.com")
iex> conn.assigns
%{}
iex> conn = Req.Conn.assign(conn, a: 1, b: 2)
iex> conn.assigns
%{a: 1, b: 2}

assign(conn, key, value)
(since 0.6.0)

@spec assign(t(), key :: atom(), value :: term()) :: t()

Assigns key/value to the conn.

Examples

iex> conn = Req.conn("http://example.com")
iex> conn.assigns
%{}
iex> conn = Req.Conn.assign(conn, :a, 1)
iex> conn.assigns
%{a: 1}

assign_new(conn, assigns)
(since 0.6.0)

@spec assign_new(t(), assigns :: keyword() | map()) :: t()

Assigns multiple keys/values to the conn unless they are already set.

Examples

iex> conn = Req.conn("http://example.com")
iex> conn.assigns
%{}
iex> conn = Req.Conn.assign_new(conn, a: 1)
iex> conn.assigns
%{a: 1}
iex> conn = Req.Conn.assign_new(conn, a: 2, b: 2)
iex> conn.assigns
%{a: 1, b: 2}

assign_new(conn, key, value)
(since 0.6.0)

@spec assign_new(t(), key :: atom(), value :: term()) :: t()

Assigns key/value to the conn unless key is already set.

Examples

iex> conn = Req.conn("http://example.com")
iex> conn.assigns
%{}
iex> conn = Req.Conn.assign_new(conn, :a, 1)
iex> conn.assigns
%{a: 1}
iex> conn = Req.Conn.assign_new(conn, :a, 2)
iex> conn.assigns
%{a: 1}

update_assign(conn, key, fun)
(since 0.6.0)

@spec update_assign(t(), key :: atom(), (term() -> term())) :: t()

Updates assign key with the given function.

Raises if the key is not set.

See also update_assign/4.

Examples

iex> conn = %Req.Conn{assigns: %{a: 1}}
iex> conn = Req.Conn.update_assign(conn, :a, & &1 * 2)
iex> conn.assigns
%{a: 2}

iex> conn = %Req.Conn{assigns: %{a: 1}}
iex> Req.Conn.update_assign(conn, :b, & &1 * 2)
** (KeyError) key :b not found in:
...

update_assign(conn, key, default, fun)
(since 0.6.0)

@spec update_assign(t(), key :: atom(), default :: term(), (term() -> term())) :: t()

Updates assign key with the given function or default.

If key is present in assigns then the existing value is passed to fun and its result is used as the updated value of key. If key is not present in assigns, default is inserted as the value of key. The default value will not be passed through the update function.

See also update_assign/3.

Examples

iex> conn = %Req.Conn{assigns: %{a: 10}}
iex> conn = Req.Conn.update_assign(conn, :a, 1, & &1 + 1)
iex> conn = Req.Conn.update_assign(conn, :b, 1, & &1 + 1)
iex> conn.assigns
%{a: 11, b: 1}

Functions

conn(conn, options \\ [])
(since 0.6.0)

@spec conn(conn :: t() | URI.t() | String.t(), options :: keyword()) :: t()

Builds a Req.Conn.

Examples

iex> Req.conn("http://example.com")
Req.conn(
  "http://example.com",
  method: "GET",
  headers: [],
  body: nil
)

iex> conn = Req.conn("https://example.com/path?x=1", method: :post)
iex> conn.host
"example.com"
iex> conn.method
"POST"

iex> conn = Req.conn("https://example.com", headers: [{"accept", "application/json"}])
iex> conn.req_headers["accept"]
"application/json"