Req.Plug.Wrapper behaviour (req v0.6.0-dev)

Behaviour for plugs that wrap the rest of the pipeline.

A wrapper receives the connection, the validated options bag, and a continuation function representing the rest of the pipeline. Wrappers may invoke the continuation zero, one, or many times.

Example

defmodule MyTime do
  @behaviour Req.Plug.Wrapper

  @impl true
  def wrap(conn, _options, next) do
    start = System.monotonic_time()
    conn = next.(conn)
    elapsed = System.monotonic_time() - start
    IO.inspect(elapsed, label: "elapsed")
    conn
  end
end

Callbacks

options()
(since 0.6.0) (optional)

@callback options() :: %Req.Options{options: term()}

wrap(t, keyword, function)
(since 0.6.0)

@callback wrap(
  Req.Conn.t(),
  keyword(),
  (Req.Conn.t() -> Req.Conn.t() | {:error, Req.Conn.t(), Exception.t()})
) :: Req.Conn.t() | {:error, Req.Conn.t(), Exception.t()}