Req.Response (req v0.3.0-dev)

The response struct.

Fields:

:status - the HTTP status code

:headers - the HTTP response headers

:body - the HTTP response body

:private - a map reserved for libraries and frameworks to use. Prefix the keys with the name of your project to avoid any future conflicts. Only accepts atom/0 keys.

Types

t()

@type t() :: %Req.Response{
  body: binary() | term(),
  headers: [{binary(), binary()}],
  private: map(),
  status: non_neg_integer()
}

Functions

get_header(response, key)

@spec get_header(t(), binary()) :: [binary()]

Returns the values of the header specified by key.

Examples

iex> Req.Response.get_header(response, "content-type")
["application/json"]

get_private(response, key, default \\ nil)

Gets the value for a specific private key.

put_header(response, key, value)

@spec put_header(t(), binary(), binary()) :: t()

Adds a new response header (key) if not present, otherwise replaces the previous value of that header with value.

Examples

iex> Req.Response.put_header(response, "content-type", "application/json").headers
[{"content-type", "application/json"}]

put_private(response, key, value)

Assigns a private key to value.