Req.Steps (req v0.3.0-dev)

The collection of built-in steps.

Req is composed of three main pieces:

Req - the high-level API

Req.Request - the low-level API and the request struct

Req.Steps - the collection of built-in steps (you're here!)

Request steps

auth(request)

Sets request authentication.

auth can be one of:

{username, password} - uses Basic HTTP authentication

{:bearer, token} - uses Bearer HTTP authentication

:netrc - load credentials from .netrc at path specified in NETRC environment variable; if NETRC is not set, load .netrc in user's home directory

{:netrc, path} - load credentials from path

Examples

iex> Req.get!("https://httpbin.org/basic-auth/foo/bar", auth: {"bad", "bad"}).status
401
iex> Req.get!("https://httpbin.org/basic-auth/foo/bar", auth: {"foo", "bar"}).status
200

iex> Req.get!("https://httpbin.org/bearer", auth: {:bearer, ""}).status
401
iex> Req.get!("https://httpbin.org/bearer", auth: {:bearer, "foo"}).status
200

iex> System.put_env("NETRC", "./test/my_netrc")
iex> Req.get!("https://httpbin.org/basic-auth/foo/bar", auth: :netrc).status
200

iex> Req.get!("https://httpbin.org/basic-auth/foo/bar", auth: {:netrc, "./test/my_netrc"}).status
200

cache(request)

Performs HTTP caching using if-modified-since header.

Only successful (200 OK) responses are cached.

This step also prepends a response step that loads and writes the cache. Be careful when prepending other response steps, make sure the cache is loaded/written as soon as possible.

Options

:cache - if true, performs caching. Defaults to false.

:cache_dir - the directory to store the cache, defaults to <user_cache_dir>/req (see: :filename.basedir/3)

Examples

iex> url = "https://elixir-lang.org"
iex> response1 = Req.get!(url, cache: true)
iex> response2 = Req.get!(url, cache: true)
iex> response1 == response2
true

compress_body(request)

Compresses the request body.

Request Options

:compress_body - if set to true, compresses the request body using gzip. Defaults to false.

compressed(request)

Asks the server to return compressed response.

Supported formats:

gzip

deflate

br (if brotli is installed)

Request Options

:compressed - if set to true, sets the accept-encoding header with compression algorithms that Req supports. Defaults to true.

Examples

Req automatically decompresses response body (decompress_body/1 step) so let's disable that by passing raw: true.

By default, we ask the server to send compressed response. Let's look at the headers and the raw body. Notice the body starts with <<31, 139>> (<<0x1F, 0x8B>>), the "magic bytes" for gzip:

iex> response = Req.get!("https://elixir-lang.org", raw: true)
iex> response.headers |> List.keyfind("content-encoding", 0)
{"content-encoding", "gzip"}
iex> response.body |> binary_part(0, 2)
<<31, 139>>

Now, let's pass compressed: false and notice the raw body was not compressed:

iex> response = Req.get!("https://elixir-lang.org", raw: true, compressed: false)
iex> response.headers |> List.keyfind("content-encoding", 0)
nil
iex> response.body |> binary_part(0, 15)
"<!DOCTYPE html>"

If the brotli package is installed, Brotli compression will also be requested:

Mix.install([
  :req,
  {:brotli, "~> 0.3.0"}
])

response = Req.get!("https://httpbin.org/brotli")
response.body["headers"]["Accept-Encoding"]
#=> "br, gzip, deflate"

encode_body(request)

Encodes the request body.

Request Options

:form - if set, encodes the request body as form data (using URI.encode_query/1).

:json - if set, encodes the request body as JSON (using Jason.encode_to_iodata!/1).

Examples

iex> Req.post!("https://httpbin.org/anything", form: [x: 1]).body["form"]
%{"x" => "1"}

iex> Req.post!("https://httpbin.org/post", json: %{x: 2}).body["json"]
%{"x" => 2}

put_base_url(request)

Sets base URL for all requests.

Request Options

:base_url - if set, the request URL is prepended with this base URL. If request URL contains a scheme, base URL is ignored.

Examples

iex> req = Req.new(base_url: "https://httpbin.org")
iex> Req.get!(req, url: "/status/200").status
200
iex> Req.get!(req, url: "/status/201").status
201

put_params(request)

Adds params to request query string.

Request Options

:params - params to add to the request query string. Defaults to [].

Examples

iex> Req.get!("https://httpbin.org/anything/query", params: [x: 1, y: 2]).body["args"]
%{"x" => "1", "y" => "2"}

put_plug(request)

Runs the request against a plug instead of over the network.

Request Options

:plug - if set, the plug to run the request against.

Examples

This step is particularly useful to test plugs:

defmodule Echo do
  def call(conn, _) do
    "/" <> path = conn.request_path
    Plug.Conn.send_resp(conn, 200, path)
  end
end

test "echo" do
  assert Req.get!("http:///hello", plug: Echo).body == "hello"
end

Here is the same example but with plug as an anonymous function:

test "echo" do
  echo = fn conn ->
    "/" <> path = conn.request_path
    Plug.Conn.send_resp(conn, 200, path)
  end

  assert Req.get!("http:///hello", plug: echo).body == "hello"
end

Here is another example, let's run the request against Plug.Static pointed to the Req's source code and fetch the README:

iex> resp = Req.get!("http:///README.md", plug: {Plug.Static, at: "/", from: "."})
iex> resp.body =~ "Req is a batteries-included HTTP client for Elixir."
true

put_range(request)

Sets the "Range" request header.

Request Options

:range - can be one of the following:

a string - returned as is

a first..last range - converted to "bytes=<first>-<last>"

Examples

iex> response = Req.get!("https://httpbin.org/range/100", range: 0..3)
iex> response.status
206
iex> response.body
"abcd"
iex> List.keyfind(response.headers, "content-range", 0)
{"content-range", "bytes 0-3/100"}

put_user_agent(request)

Sets the user-agent header.

Request Options

:user_agent - sets the user-agent header. Defaults to "req/0.3.0-dev".

Examples

iex> Req.get!("https://httpbin.org/user-agent").body
%{"user-agent" => "req/0.3.0-dev"}

iex> Req.get!("https://httpbin.org/user-agent", user_agent: "foo").body
%{"user-agent" => "foo"}

run_finch(request)

Runs the request using Finch.

This is the default Req adapter. See "Adapter" section in the Req.Request module documentation for more information on adapters.

Request Options

:finch - the name of the Finch pool. Defaults to a pool automatically started by Req. The default pool uses HTTP/1 although that may change in the future. This option takes precedence over :http1 and :http2 options mentioned below.

:finch_options - options passed down to Finch when making the request, defaults to []. See Finch.request/3 for a list of available options.

:http1 - if true, uses an HTTP/1 pool automatically started by Req.

:http2 - if true, uses an HTTP/2 pool automatically started by Req.

:unix_socket - if set, connect through the given UNIX domain socket

Examples

Custom :receive_timeout:

iex> Req.get!(url: url, finch_options: [receive_timeout: 1000])

Connecting through UNIX socket:

iex> Req.get!("http:///v1.41/_ping", unix_socket: "/var/run/docker.sock").body
"OK"

Response steps

decode_body(request_response)

Decodes response body based on the detected format.

Supported formats:

Format	Decoder

json	Jason.decode!/1

gzip	:zlib.gunzip/1

tar, tgz	:erl_tar.extract/2

zip	:zip.unzip/2

csv	NimbleCSV.RFC4180.parse_string/2 (if nimble_csv is installed)

Request Options

:decode_body - if set to false, disables automatic response body decoding. Defaults to true.

Examples

iex> response = Req.get!("https://httpbin.org/gzip")
...> response.body["gzipped"]
true

iex> response = Req.get!("https://httpbin.org/json")
...> response.body["slideshow"]["title"]
"Sample Slide Show"

decompress_body(request_response)

Decompresses the response body based on the content-encoding header.

Supported formats:

Format	Decoder

gzip, x-gzip	:zlib.gunzip/1

zip	:zlib.unzip/1

br	:brotli.decode/1 (if brotli is installed)

Examples

iex> response = Req.get!("https://httpbin.org/gzip")
iex> response.headers |> List.keyfind("content-encoding", 0)
{"content-encoding", "gzip"}
iex> response.body["gzipped"]
true

If the brotli package is installed, Brotli is also supported:

Mix.install([
  :req,
  {:brotli, "~> 0.3.0"}
])

response = Req.get!("https://httpbin.org/brotli")
response.headers |> List.keyfind("content-encoding", 0)
#=> {"content-encoding", "br"}
response.body["brotli"]
#=> true

follow_redirects(request_response)

Follows redirects.

The original request method may be changed to GET depending on the status code:

Code	Method handling

301, 302, 303	Changed to GET

307, 308	Method not changed

Request Options

:follow_redirects - if set to false, disables automatic response redirects. Defaults to true.

:location_trusted - by default, authorization credentials are only sent on redirects with the same host, scheme and port. If :location_trusted is set to true, credentials will be sent to any host.

:max_redirects - the maximum number of redirects, defaults to 10. If the limit is reached, an erorr is raised.

Examples

iex> Req.get!("http://api.github.com").status
# 23:24:11.670 [debug]  follow_redirects: redirecting to https://api.github.com/
200

iex> Req.get!("https://httpbin.org/redirect/4", max_redirects: 3)
# 23:07:59.570 [debug] follow_redirects: redirecting to /relative-redirect/3
# 23:08:00.068 [debug] follow_redirects: redirecting to /relative-redirect/2
# 23:08:00.206 [debug] follow_redirects: redirecting to /relative-redirect/1
** (RuntimeError) too many redirects (3)

handle_http_errors(request_response)

Handles HTTP 4xx/5xx error responses.

Request Options

:http_errors - how to handle HTTP 4xx/5xx error responses. Can be one of the following:

:return (default) - return the response

:raise - raise an error

Examples

iex> Req.get!("https://httpbin.org/status/404").status
404

iex> Req.get!("https://httpbin.org/status/404", http_errors: :raise)
** (RuntimeError) The requested URL returned error: 404
Response body: ""

output(arg)

Writes the response body to a file.

After the output file is written, the response body is set to "".

Request Options

:output - if set, writes the response body to a file. Can be one of:

path - writes to the given path

:remote_name - uses the remote name as the filename in the current working directory

Examples

iex> Req.get!("https://elixir-lang.org/index.html", output: "/tmp/elixir_home.html")
iex> File.exists?("/tmp/elixir_home.html")
true

iex> Req.get!("https://elixir-lang.org/blog/index.html", output: :remote_name)
iex> File.exists?("index.html")
true

Error steps

retry(request_response_or_error)

Retries a request in face of errors.

This function can be used as either or both response and error step.

Request Options

:retry - can be one of the following:

:safe (default) - retry GET/HEAD requests on HTTP 408/5xx responses or exceptions

:always - always retry

:never - never retry

fun - a 1-arity function that accepts either a Req.Response or an exception struct and returns boolean whether to retry

:retry_delay - sleep this number of milliseconds before making another attempt, defaults to 2000

:max_retries - maximum number of retry attempts, defaults to 2 (for a total of 3 requests to the server, including the initial one.)

Examples

With default options:

iex> Req.get!("https://httpbin.org/status/500,200").status
# 19:02:08.463 [error] retry: got response with status 500, will retry in 2000ms, 2 attempts left
# 19:02:10.710 [error] retry: got response with status 500, will retry in 2000ms, 1 attempt left
200

With custom options:

iex> Req.get!("http://localhost:9999", retry_delay: 100, max_retries: 3)
# 17:00:38.371 [error] retry: got exception, will retry in 100ms, 3 attempts left
# 17:00:38.371 [error] ** (Mint.TransportError) connection refused
# 17:00:38.473 [error] retry: got exception, will retry in 100ms, 2 attempts left
# 17:00:38.473 [error] ** (Mint.TransportError) connection refused
# 17:00:38.575 [error] retry: got exception, will retry in 100ms, 1 attempt left
# 17:00:38.575 [error] ** (Mint.TransportError) connection refused
** (Mint.TransportError) connection refused