Req (req v0.1.0-dev)

An HTTP client with a focus on composability, built on top of Finch.

This is a work in progress!

Features

Extensibility via request, response, and error steps

Automatic body decompression (via decompress/2 step)

Automatic body encoding and decoding (via encode/1 and decode/2 steps)

Encode params as query string (via params/2 step)

Basic authentication (via auth/2 step)

Follows redirects (via follow_redirects/2 step)

Retries on errors (via retry/3 step)

Basic HTTP caching (via if_modified_since/2 step)

High-level API

Req.get!("https://api.github.com/repos/elixir-lang/elixir").body["description"]
#=> "Elixir is a dynamic, functional language designed for building scalable and maintainable applications"

Low-level API

Under the hood, Req works by passing a request through a series of steps.

The request struct, %Req.Request{}, initially contains data like HTTP method and request headers. You can also add request, response, and error steps to it.

Request steps are used to refine the data that will be sent to the server.

After making the actual HTTP request, we'll either get a HTTP response or an error. The request, along with the response or error, will go through response or error steps, respectively.

Nothing is actually executed until we run the pipeline with Req.run/1.

Example:

Req.build(:get, "https://api.github.com/repos/elixir-lang/elixir")
|> Req.add_request_steps([
  &Req.default_headers/1
])
|> Req.add_response_steps([
  &Req.decode/2
])
|> Req.run()
#=> {:ok, %{body: %{"description" => "Elixir is a dynamic," <> ...}, ...}, ...}

The high-level API shown before:

Req.get!("https://api.github.com/repos/elixir-lang/elixir")

is equivalent to this composition of lower-level API functions:

Req.build(:get, "https://api.github.com/repos/elixir-lang/elixir")
|> Req.add_default_steps()
|> Req.run!()

(See Req.build/3, Req.add_default_steps/2, and Req.run!/1 for more information.)

We can also build more complex flows like returning a response from a request step or an error from a response step. We will explore those next.

Request steps

A request step is a function that accepts a request and returns one of the following:

A request

A {request, response_or_error} tuple. In that case no further request steps are executed and the return value goes through response or error steps

Examples:

def default_headers(request) do
  update_in(request.headers, &[{"user-agent", "req/0.1.0-dev"} | &1])
end

def read_from_cache(request) do
  case ResponseCache.fetch(request) do
    {:ok, response} -> {request, response}
    :error -> request
  end
end

Response and error steps

A response step is a function that accepts a request and a response and returns one of the following:

A {request, response} tuple

A {request, exception} tuple. In that case, no further response steps are executed but the exception goes through error steps

Similarly, an error step is a function that accepts a request and an exception and returns one of the following:

A {request, exception} tuple

A {request, response} tuple. In that case, no further error steps are executed but the response goes through response steps

Examples:

def decode(request, response) do
  case List.keyfind(response.headers, "content-type", 0) do
    {_, "application/json" <> _} ->
      {request, update_in(response.body, &Jason.decode!/1)}

    _ ->
      {request, response}
  end
end

def log_error(request, exception) do
  Logger.error(["#{request.method} #{request.url}: ", Exception.message(exception)])
  {request, exception}
end

Halting

Any step can call Req.Request.halt/1 to halt the pipeline. This will prevent any further steps from being invoked.

Examples:

def circuit_breaker(request) do
  if CircuitBreaker.open?() do
    {Req.Request.halt(request), RuntimeError.exception("circuit breaker is open")}
  else
    request
  end
end

High-level API

get!(uri, options \\ [])

Makes a GET request.

See request/3 for a list of supported options.

post!(uri, body, options \\ [])

Makes a POST request.

See request/3 for a list of supported options.

request(method, uri, options \\ [])

Makes an HTTP request.

Options

:header - request headers, defaults to []

:body - request body, defaults to ""

:finch - Finch pool to use, defaults to Req.Finch which is automatically started by the application. See Finch module documentation for more information on starting pools.

:finch_options - Options passed down to Finch when making the request, defaults to []. See Finch.request/3 for more information.

The options are passed down to add_default_steps/2, see its documentation for more information how they are being used.

request!(method, uri, options \\ [])

Makes an HTTP request and returns a response or raises an error.

See request/3 for more information.

Low-level API

add_default_steps(request, options \\ [])

Adds steps that should be reasonable defaults for most users.

Request steps

normalize_headers/1

default_headers/1

encode/1

&auth(&1, options[:auth]) (if options[:auth] is set)

&params(&1, options[:params]) (if options[:params] is set)

Response steps

&retry(&1, &2, options[:retry]) (if options[:retry] is set and is a keywords list or an atom true)

follow_redirects/2

decompress/2

decode/2

Error steps

&retry(&1, &2, options[:retry]) (if options[:retry] is set and is a keywords list or an atom true)

Options

:auth - if set, adds the auth/2 step

:params - if set, adds the params/2 step

:cache - if set to true, adds if_modified_since/2 step

:raw if set to true, skips decompress/2 and decode/2 steps

:retry - if set, adds the retry/3 step to response and error steps

add_error_steps(request, steps)

Adds error steps.

add_request_steps(request, steps)

Adds request steps.

add_response_steps(request, steps)

Adds response steps.

build(method, uri, options \\ [])

Builds a request pipeline.

Options

:header - request headers, defaults to []

:body - request body, defaults to ""

:finch - Finch pool to use, defaults to Req.Finch which is automatically started by the application. See Finch module documentation for more information on starting pools.

:finch_options - Options passed down to Finch when making the request, defaults to []. See Finch.request/3 for more information.

run(request)

Runs a request pipeline.

Returns {:ok, response} or {:error, exception}.

run!(request)

Runs a request pipeline and returns a response or raises an error.

See run/1 for more information.

Request steps

auth(request, auth)

Sets request authentication.

auth can be one of:

{username, password} - uses Basic HTTP authentication

Examples

iex> Req.get!("https://httpbin.org/basic-auth/foo/bar", auth: {"bad", "bad"}).status
401
iex> Req.get!("https://httpbin.org/basic-auth/foo/bar", auth: {"foo", "bar"}).status
200

default_headers(request)

Adds common request headers.

Currently the following headers are added:

"user-agent" - "req/0.1.0-dev"

"accept-encoding" - "gzip"

encode(request)

Encodes the request body based on its shape.

If body is of the following shape, it's encoded and its content-type set accordingly. Otherwise it's unchanged.

Shape	Encoder	Content-Type

{:form, data}	URI.encode_query/1	"application/x-www-form-urlencoded"

{:json, data}	Jason.encode_to_iodata!/1	"application/json"

Examples

iex> Req.post!("https://httpbin.org/post", {:form, comments: "hello!"}).body["form"]
%{"comments" => "hello!"}

if_modified_since(request, options \\ [])

Handles HTTP cache using if-modified-since header.

Only successful (200 OK) responses are cached.

This step also prepends a response step that loads and writes the cache. Be careful when prepending other response steps, make sure the cache is loaded/written as soon as possible.

Options

:dir - the directory to store the cache, defaults to <user_cache_dir>/req (see: :filename.basedir/3)

Examples

iex> url = "https://hexdocs.pm/elixir/Kernel.html"
iex> response = Req.get!(url, cache: true)
%{
  status: 200,
  headers: [
    {"date", "Fri, 16 Apr 2021 10:09:56 GMT"},
    ...
  ],
  ...
}
iex> Req.get!(url, cache: true) == response
true

normalize_headers(request)

Normalizes request headers.

Turns atom header names into strings, e.g.: :user_agent becomes "user-agent". Non-atom names are returned as is.

Examples

iex> Req.get!("https://httpbin.org/user-agent", headers: [user_agent: "my_agent"]).body
%{"user-agent" => "my_agent"}

params(request, params)

Adds params to request query string.

Examples

iex> Req.get!("https://httpbin.org/anything/query", params: [x: "1", y: "2"]).body["args"]
%{"x" => "1", "y" => "2"}

Response steps

decode(request, response)

Decodes response body based on the detected format.

Supported formats:

Format	Decoder

json	Jason.decode!/1

gzip	:zlib.gunzip/1

tar	:erl_tar.extract/2

zip	:zip.unzip/2

csv	NimbleCSV.RFC4180.parse_string/2 (if NimbleCSV is installed)

Examples

iex> Req.get!("https://hex.pm/api/packages/finch").body["meta"]
%{
  "description" => "An HTTP client focused on performance.",
  "licenses" => ["MIT"],
  "links" => %{"GitHub" => "https://github.com/keathley/finch"},
  ...
}

decompress(request, response)

Decompresses the response body based on the content-encoding header.

Examples

iex> response = Req.get!("https://httpbin.org/gzip")
iex> response.headers
[
  {"content-encoding", "gzip"},
  {"content-type", "application/json"},
  ...
]
iex> response.body
%{
  "gzipped" => true,
  ...
}

follow_redirects(request, response)

Follows redirects.

Examples

iex> Req.get!("http://api.github.com").status
# 23:24:11.670 [debug]  Req.follow_redirects/2: Redirecting to https://api.github.com/
200

Error steps

retry(request, response_or_exception, options \\ [])

Retries a request in face of errors.

This function can be used as either or both response and error step. It retries a request that resulted in:

a response with status 5xx

an exception

Options

:delay - sleep this number of milliseconds before making another attempt, defaults to 2000

:max_attempts - maximum number of attempts, defaults to 2

Examples

With default options:

iex> Req.get!("https://httpbin.org/status/500,200", retry: true).status
# 19:02:08.463 [error] Req.retry/3: Got response with status 500. Will retry in 2000ms, 2 attempts left
# 19:02:10.710 [error] Req.retry/3: Got response with status 500. Will retry in 2000ms, 1 attempt left
200

With custom options:

iex> Req.get!("http://localhost:9999", retry: [delay: 100, max_attempts: 3])
# 17:00:38.371 [error] Req.retry/3: Got exception. Will retry in 100ms, 3 attempts left
# 17:00:38.371 [error] ** (Mint.TransportError) connection refused
# 17:00:38.473 [error] Req.retry/3: Got exception. Will retry in 100ms, 2 attempts left
# 17:00:38.473 [error] ** (Mint.TransportError) connection refused
# 17:00:38.575 [error] Req.retry/3: Got exception. Will retry in 100ms, 1 attempt left
# 17:00:38.575 [error] ** (Mint.TransportError) connection refused
** (Mint.TransportError) connection refused